+++ Makefile
@@ -1,4 +1,4 @@
-BIOS_NAME := BOOTX64.EFI
+LOADER_NAME := BOOTX64.EFI
 KERNEL_NAME := KizunaOS.ELF
 SRCDIR := src
 OBJDIR := build
@@ -7,67 +7,66 @@ APPDIR := $(OUTDIR)/EFI/BOOT
 HOMEDIR := $(shell echo $$HOME)
 WORKDIR := $(shell pwd)
 
-BIOS_TARGET = $(APPDIR)/$(BIOS_NAME)
+LOADER_TARGET = $(APPDIR)/$(LOADER_NAME)
 KERNEL_TARGET = $(OUTDIR)/$(KERNEL_NAME)
 
 CC = clang++
-BIOS_CFLAGS = \
-	--target=x86_64-pc-win32-coff \
-	-Wall -Wextra -Wpedantic
-
-LOADER_CFAGS = \
-
-BIOS_CPPFLAGS = \
-	-MMD -MP \
-	-I $(HOMEDIR)/x86_64-elf/include -I $(HOMEDIR)/x86_64-elf/include/c++/v1 \
-	-D__ELF__ -D_LIBCPP_HAS_NO_THREADS \
+LOADER_CFLAGS = \
+	-I $(HOMEDIR)/x86_64-elf/include \
 	--target=x86_64-pc-win32-coff \
 	-fno-stack-protector -fno-exceptions -fshort-wchar \
 	-nostdlibinc -mno-red-zone \
-	-Wall -Wextra -Wpedantic -Qunused-arguments -Wno-keyword-macro -Wno-char-subscripts -Wno-int-to-pointer-cast\
-	-Wno-c99-extensions -Wno-unused-parameter -Wno-unused-variable -Wno-writable-strings -Wno-macro-redefined\
+	-Wall -Wextra -Wpedantic
+
+LOADER_CPPFLAGS = \
+	-I $(HOMEDIR)/x86_64-elf/include \
+	-D_LIBCPP_HAS_NO_THREADS \
+	--target=x86_64-pc-win32-coff \
+	-fno-stack-protector -fno-exceptions -fshort-wchar \
+	-nostdlibinc -nostdlib -mno-red-zone \
+	-Wall -Wextra -Wpedantic -Qunused-arguments -Wno-keyword-macro -Wno-char-subscripts -Wno-int-to-pointer-cast \
+	-Wno-c99-extensions -Wno-unused-parameter -Wno-unused-variable -Wno-writable-strings -Wno-macro-redefined \
 	-fno-builtin \
 	-std=c++17
 
 KERNEL_CPPFLAGS = \
-	-MMD -MP \
 	-I $(HOMEDIR)/x86_64-elf/include -I $(HOMEDIR)/x86_64-elf/include/c++/v1 \
 	-D__ELF__ -D_LIBCPP_HAS_NO_THREADS \
 	--target=x86_64-unknown-none-elf \
 	-fno-stack-protector -fno-exceptions -fshort-wchar \
-	-nostdlibinc -mno-red-zone \
+	-nostdlibinc -ffreestanding -mno-red-zone \
 	-Wall -Wextra -Wpedantic -Qunused-arguments -Wno-keyword-macro -Wno-char-subscripts -Wno-int-to-pointer-cast \
 	-Wno-c99-extensions -Wno-unused-parameter -Wno-unused-variable -Wno-writable-strings -Wno-macro-redefined -Wno-sign-compare\
 	-fno-builtin \
-	-fPIC -Wl,-pie \
+	-g \
 	-std=c++17
 
-LD_LINK = /usr/bin/lld-link-6.0
+LD_LINK = /usr/bin/lld-link-10
 LD_LLD = /usr/bin/ld.lld
-BIOS_LDFLAGS = \
+LOADER_LDFLAGS = \
 	-subsystem:efi_application -nodefaultlib \
 	-entry:efi_main
 
 KERNEL_LDFLAGS = \
-	-T kernel.ld -static
+	-T kernel.ld -L$(HOMEDIR)/x86_64-elf/lib -lc++ -lc++abi -lm -lc -static
 
 QEMU = qemu-system-x86_64
 OVMF = ovmf/bios64.bin
 QEMUflags = \
 	-bios $(OVMF) -drive format=raw,file=fat:rw:$(OUTDIR)
 
-BIOS_SRCS = \
+LOADER_SRCS = \
 	boot_loader.cpp efi_main.cpp efi.cpp efi_kernel_loader.cpp loader_asm.s std_func.cpp graphics.cpp
 KERNEL_SRCS = \
-	kernel.cpp kernel_asm.s std_func.cpp graphics.cpp
+	kernel.cpp kernel_asm.s std_func.cpp graphics.cpp libc_support.c
 
 SRCS = $(wildcard $(SRCDIR)/*.cpp)
-BIOS_OBJS := $(addprefix $(OBJDIR)/,$(addsuffix .o, $(basename $(notdir $(BIOS_SRCS)))))
+LOADER_OBJS := $(addprefix $(OBJDIR)/,$(addsuffix .o, $(basename $(notdir $(LOADER_SRCS)))))
 KERNEL_OBJS := $(addprefix $(OBJDIR)/,$(addsuffix .elf.o, $(basename $(KERNEL_SRCS))))
 DEPS := $(addprefix $(OBJDIR)/,$(patsubst $(SRCDIR)/%.cpp,%.d,$(SRCS)))
 
 .PHONY: all clean
-all: Makefile $(APPDIR) $(OBJDIR) $(TOOLS) $(BIOS_TARGET) $(KERNEL_TARGET)
+all: Makefile $(APPDIR) $(OBJDIR) $(TOOLS) $(LOADER_TARGET) $(KERNEL_TARGET)
 
 $(APPDIR):
 	mkdir -p $(APPDIR)
@@ -76,22 +75,25 @@ $(OBJDIR):
 	mkdir -p $(OBJDIR)
 
 $(OBJDIR)/%.o:$(SRCDIR)/%.s
-	$(CC) $(BIOS_CFLAGS) -o $@ -c $<
+	$(CC) $(LOADER_CFLAGS) -o $@ -c $<
 
 $(OBJDIR)/%.o:$(SRCDIR)/%.cpp
-	$(CC) $(BIOS_CPPFLAGS) -o $@ -c $<
+	$(CC) $(LOADER_CPPFLAGS) -o $@ -c $<
 
 $(OBJDIR)/%.elf.o:$(SRCDIR)/%.s
+	$(CC) $(KERNEL_CFLAGS) -o $@ -c $<
+
+$(OBJDIR)/%.elf.o:$(SRCDIR)/%.c
 	$(CC) $(KERNEL_CFLAGS) -o $@ -c $<
 
 $(OBJDIR)/%.elf.o:$(SRCDIR)/%.cpp
 	$(CC) $(KERNEL_CPPFLAGS) -o $@ -c $<
 
-$(BIOS_TARGET): $(BIOS_OBJS)
-	$(LD_LINK) $(BIOS_LDFLAGS) -out:$(BIOS_TARGET) $(BIOS_OBJS)
+$(LOADER_TARGET): $(LOADER_OBJS)
+	$(LD_LINK) $(LOADER_LDFLAGS) -out:$(LOADER_TARGET) $(LOADER_OBJS)
 
 $(KERNEL_TARGET): $(KERNEL_OBJS) kernel.ld
-	$(LD_LLD) $(KERNEL_LDFLAGS) -o $(KERNEL_TARGET) $(KERNEL_OBJS) --pic-executable
+	$(LD_LLD) $(KERNEL_LDFLAGS) -o $(KERNEL_TARGET) $(KERNEL_OBJS)
 
 $(TOOLS):
 	mkdir -p work
@@ -140,7 +142,7 @@ $(TOOLS):
 
 tools: Makefile $(TOOLS)
 
-bios: Makefile $(OBJDIR) $(BIOS_TARGET)
+loader: Makefile $(OBJDIR) $(LOADER_TARGET)
 
 run: $(TARGET)
 	$(QEMU) $(QEMUflags)
